Ensure "--args" is last argument of gdb command
To debug an application+command line arguments "--args" is used as last command line argument of gdbgui.
Everything after "--args" must be passed then to gdb also as LAST command line argument.
This was not the case, but it is now ensured. By default gdb is now used in any case with "--args".